+++ previous-versions/files/src/sdk/sdk.ts
@@ -53,8 +53,9 @@ export class SDKConfiguration {
     serverDefaults: any;
     language = "typescript";
     openapiDocVersion = "3.0.0";
-    sdkVersion = "0.35.0";
-    genVersion = "2.109.3";
+    sdkVersion = "0.36.0";
+    genVersion = "2.144.7";
+    userAgent = "speakeasy-sdk/typescript 0.36.0 2.144.7 3.0.0 @codat/files";
     retryConfig?: utils.RetryConfig;
     public constructor(init?: Partial<SDKConfiguration>) {
         Object.assign(this, init);